requirements.txt
Flask>=2.0
groq>=0.5.0
openai>=1.0.0
anthropic>=0.20.0
python-dotenv>=1.0.0
typing_extensions>=4.0.0
google-genai>=0.1.0